general cleaning and finishing touches
